requirements.txt
Django==2.0.7  
djangorestframework==3.10.2 
pytz==2019.2
django-request-logging==0.7.0


